+++ clients/python/src/osparc/_api_solvers_api.py
