+++ clients/python/src/osparc/_api_studies_api.py
